Enable breadcrumb context tracking with DRED
