+++ src/TerraFX.Interop.Windows/TerraFX.Interop.Windows.projitems
@@ -189,7 +189,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12CommandQueue.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12DescriptorHeap.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12Device.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12DeviceRemovedExtendedDataSettings.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12DeviceRemovedExtendedDataSettings1.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12Fence.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12GraphicsCommandList.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DirectX\um\d3d12\ID3D12Heap.cs" />
